Enforce the order of `import` statements.
This makes the files more readable, allowing for an easy scanning of what is imported and makes it faster to find a specific import.

They will be grouped by types, then sorted alphabeticaly in each group.

The groups will be:
- Internal imports (NodeJS)
- React / ReactDOM / PropTypes
- React Intl
- Immutable / Redux
- absolute `mastodon/` imports
- relative imports

I chose to group the common imports in their own groups, so they can be more easily scanned.

This is auto-fixed by ESLint, so it does not bring any developer overhead.

I chose to only apply it to Typescript files as there are only a few of them for now, but this could be applied to all JS files if wanted.