Add an optional metric exporter
I tried to only require the gem when the feature is enabled.

When enabled, it will report some metrics, and additional env vars can enable some others (per-request / per job metrics) that might not be needed by everyone, as OTEL can provide better information about those.

By default, this expects a `prometheus_exporter` running locally. As an alternative, `MASTODON_PROMETHEUS_EXPORTER_LOCAL` can be configured to expose the metrics from the current process, to avoid the overhead. Note that when running Puma, you can not run this option as it spawns multiple processes, but it can be useful for Sidekiq in container environments.